Add associated *shapes* alongside associated type factories
This is exposed as a separate attribute property because factories and shapes are distinct and attribute construction has very limited syntax. But internally, we have just one associated type model with flags that indicate what the requirements are.

This new functionality is exposed on `TypeShapeAttribute`, `TypeShapeExtensionAttribute` and `AssociatedTypeAttributeAttribute`.

The acceptance test for associated types has been updated as well. The case may not look particularly compelling, so in short the case is that an associated *type* may be a custom converter and an associated *shape* may be a convertible type that the custom converter uses as a part of its serialization.

Closes #146