Replace associated type factories with more flexible associated *shapes* (#155)
* Add associated *shapes* alongside associated type factories

This is exposed as a separate attribute property because factories and shapes are distinct and attribute construction has very limited syntax. But internally, we have just one associated type model with flags that indicate what the requirements are.

This new functionality is exposed on `TypeShapeAttribute`, `TypeShapeExtensionAttribute` and `AssociatedTypeAttributeAttribute`.

The acceptance test for associated types has been updated as well. The case may not look particularly compelling, so in short the case is that an associated *type* may be a custom converter and an associated *shape* may be a convertible type that the custom converter uses as a part of its serialization.

Closes #146

* Replace default constructor factories for partial shapes

* Rename `TypeShapeDepth` to `TypeShapeRequirements`

* Rename `TypeShapeRequirements.All` to `Full`

Also update other doc comments.

* Switch to throwing instead of returning empty from partial shapes

* Add more tests

- Verify that we throw on attempts to retrieve missing constructors.
- Verify that partial shapes by association are made complete by references from other shapes.

* Generate empty shapes when `TypeShapeRequirements` is set to `None`

* Update  docs

* Rename more 'depth' references to `Requirements`

* Apply doc touch-ups from code review

Co-authored-by: Eirik Tsarpalis <eirik.tsarpalis@gmail.com>

---------

Co-authored-by: Eirik Tsarpalis <eirik.tsarpalis@gmail.com>